Isprime extension

+++ NumberGen.Model/NgPrime.cs
@@ -6,5 +6,5 @@ public class NgPrime
 {
     public ulong Number { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
-    public int GenerationTime { get; set; }
+    public long GenerationTime { get; set; }
 }
