+++ NumberGen/Responses/NgPrimeResponse.cs
@@ -4,5 +4,5 @@ public class NgPrimeResponse
 {
     public ulong Number { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
-    public int GenerationTime { get; set; }
+    public long GenerationTime { get; set; }
 }
